+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Final model returned from `TorchTrainStep` ("torch::train") will have best weights loaded.
 - Checkpoints are saved from `TorchTrainStep` ("torch::train") even when there is no validation loop.
+- Fixed `TorchTrainStep` ("torch::train") when `validation_split` is `None`.
 
 ## [v0.1.2](https://github.com/allenai/tango/releases/tag/v0.1.2) - 2021-10-13